Generating autotools scripts
Many popular projects using autotools, for example, GCC, GLIBC,
store the generated configuration and building scripts in the
project repository. This is necessary because the project requires
a specific version of autotools, to speed up the build, and it's
not necessary to regenerate scripts on the end user's machine,
it's enough to do this once to the project's maintainer.